refactor: Extract parallelMap as util
